[feat-metamask]-chore: delete file
